refactor: update sharedweb references to commonweb for consistency
