+++ common-web/src/main/java/pt/estga/commonweb/config/RateLimitProperties.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedweb.config;
+package pt.estga.commonweb.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
