+++ common-web/src/main/java/pt/estga/commonweb/exceptions/UnsupportedFileTypeException.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedweb.exceptions;
+package pt.estga.commonweb.exceptions;
 
 public class UnsupportedFileTypeException extends FileStorageException {
     public UnsupportedFileTypeException(String message) {
