+++ file/src/main/java/pt/estga/file/controllers/MediaController.java
@@ -76,7 +76,7 @@ public class MediaController {
         MediaFile mediaFile = mediaService.findById(id)
                 .orElseThrow(() -> {
                     log.warn("Media not found with id: {}", id);
-                    return new pt.estga.sharedweb.exceptions.FileNotFoundException("Media not found with id: " + id);
+                    return new pt.estga.commonweb.exceptions.FileNotFoundException("Media not found with id: " + id);
                 });
 
         log.info("Found media file, loading from path: {}", mediaFile.getStoragePath());
