+++ file/src/main/java/pt/estga/file/services/FileStorageAdapter.java
@@ -31,7 +31,7 @@ public class FileStorageAdapter implements FileStorageOperations {
             var entity = orchestrator.orchestrateUpload(data, originalFilename);
             return MediaFileMapper.toDto(entity);
         } catch (java.io.IOException e) {
-            throw new pt.estga.sharedweb.exceptions.FileStorageException("Failed to upload file", e);
+            throw new pt.estga.commonweb.exceptions.FileStorageException("Failed to upload file", e);
         }
     }
 
@@ -51,7 +51,7 @@ public class FileStorageAdapter implements FileStorageOperations {
                 return MediaFileMapper.toDto(entity);
             }
         } catch (java.io.IOException e) {
-            throw new pt.estga.sharedweb.exceptions.FileStorageException("Failed to commit staged file " + stagingId, e);
+            throw new pt.estga.commonweb.exceptions.FileStorageException("Failed to commit staged file " + stagingId, e);
         }
     }
