+++ file/src/test/java/pt/estga/file/controllers/MediaControllerTest.java
@@ -8,12 +8,11 @@ import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import pt.estga.file.dtos.MediaFileDto;
 import pt.estga.file.entities.MediaFile;
 import pt.estga.file.services.application.MediaService;
 import pt.estga.file.services.application.MediaVariantService;
-import pt.estga.sharedweb.exceptions.FileNotFoundException;
-import pt.estga.sharedweb.exceptions.GlobalExceptionHandler;
+import pt.estga.commonweb.exceptions.FileNotFoundException;
+import pt.estga.commonweb.exceptions.GlobalExceptionHandler;
 
 import java.util.UUID;
