+++ mark/src/main/java/pt/estga/mark/services/MarkService.java
@@ -12,7 +12,7 @@ import pt.estga.mark.entities.MarkEvidence;
 import pt.estga.mark.mappers.MarkMapper;
 import pt.estga.mark.repositories.MarkEvidenceRepository;
 import pt.estga.mark.repositories.MarkRepository;
-import pt.estga.sharedweb.exceptions.ResourceNotFoundException;
+import pt.estga.commonweb.exceptions.ResourceNotFoundException;
 
 @Service
 @RequiredArgsConstructor
