+++ monument/src/main/java/pt/estga/monument/services/MonumentService.java
@@ -9,7 +9,7 @@ import pt.estga.monument.dtos.MonumentFilter;
 import pt.estga.monument.entities.Monument;
 import pt.estga.monument.MonumentRepository;
 import pt.estga.commoninfra.jpa.SpecBuilder;
-import pt.estga.sharedweb.exceptions.ResourceNotFoundException;
+import pt.estga.commonweb.exceptions.ResourceNotFoundException;
 
 import java.util.Optional;
