+++ processing/src/main/java/pt/estga/processing/controllers/ProcessingController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import pt.estga.processing.dtos.ProcessingStatusDto;
 import pt.estga.processing.services.processing.AsyncProcessingService;
 import pt.estga.processing.services.processing.ProcessingStatusQueryService;
-import pt.estga.sharedweb.dtos.MessageResponseDto;
+import pt.estga.commonweb.dtos.MessageResponseDto;
 
 @RestController
 @RequestMapping("/api/v1/admin/processing")
